requirements.txt
python-dotenv>=1.0
streamlit>=1.37
openai>=1.40.0
tiktoken>=0.7
requests>=2.32
soundfile>=0.12
